refactor(md-ast): consolidate MarkdownFeaturesConfig + per-feature option structs
Wave 2 (#569) initially shipped a duplicate MarkdownFeaturesConfig in
zfb-md-extras (LESS rich than zfb's version — github_autolinks and
heading_marker_toc collapsed to Option<FeatureToggle> instead of typed
configs). Codex review flagged both the duplication and the shape
mismatch in pass 2.

Resolution: move the canonical definitions to zfb-md-ast (the lowest
crate; both zfb and zfb-md-extras can depend on it without a cycle).
This removes:

- zfb-md-extras's duplicate FeatureToggle / FeatureOptions /
  MarkdownFeaturesConfig — now re-exports from zfb-md-ast.
- zfb's inline FeatureToggle / FeatureOptions / MarkdownFeaturesConfig /
  GithubAutolinksConfig / CodeEnrichmentConfig / TocExportConfig /
  ImageDimensionsConfig / LinkValidationConfig / TranscludeConfig — now
  re-exports from zfb-md-ast. zfb adds a direct zfb-md-ast dep (already
  transitively present via zfb-content).
- TocConfig moved from zfb-content/src/plugins/toc.rs to zfb-md-ast so
  the heading_marker_toc field can carry the rich shape. The TocPlugin
  still lives in zfb-content and now imports TocConfig via zfb-md-ast.
  zfb-content::plugins::toc::TocConfig + zfb-content::TocConfig + zfb::
  config::TocConfig all still resolve via re-export shims.

Dep graph stays acyclic — verified via cargo metadata:
  zfb-md-ast (no zfb-* deps, bottom)
  → zfb-md-extras
  → zfb-content (also depends on zfb-md-ast directly)
  → zfb (also depends on zfb-md-ast directly)

The Pipeline::with_defaults_and_features sibling constructor introduced
by #569 now takes &zfb_md_ast::MarkdownFeaturesConfig — the SAME type
zfb::config users construct, no conversion bridge needed.
