+++ crates/zfb-md-ast/Cargo.toml
@@ -9,3 +9,4 @@ description = "zfb-md-ast: shared MdastVisitor / HastVisitor traits and HastNode
 
 [dependencies]
 markdown = { workspace = true }
+serde = { workspace = true }
