+++ crates/zfb/src/config.rs
@@ -757,10 +757,10 @@ pub struct PluginConfig {
 
 // --- markdown / GFM config -------------------------------------------------
 
-/// Re-export [`TocConfig`] from `zfb-content` so `zfb.config.ts`
-/// consumers (and the bundler / snapshot callers) reach the type without
-/// a direct `zfb-content` dependency.
-pub use zfb_content::TocConfig;
+// TocConfig now lives in `zfb-md-ast::features_config` (re-exported below
+// alongside the other feature config types). `zfb-content` still
+// re-exports it under the historical `zfb_content::TocConfig` path for
+// downstream code that imports through that crate's lib surface.
 
 /// Markdown / MDX parsing options.
 ///
@@ -834,154 +834,18 @@ pub struct MarkdownConfig {
 }
 
 // --- MarkdownFeaturesConfig and per-feature types --------------------------
-
-/// Per-feature toggle: `bool` shorthand or a feature-options object.
-///
-/// `true` enables the feature with defaults; `false` (or absent) disables
-/// it. The object form carries per-feature options (fields vary by feature
-/// and are filled in by each feature's port sub-issue — stubs today).
-///
-/// `serde(untagged)` with the `Options` variant listed FIRST ensures serde
-/// tries the object form before the bool, mirroring the `GfmFlag` pattern
-/// already used in this file.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum FeatureToggle {
-    /// Per-feature options object (fields are feature-specific).
-    Options(FeatureOptions),
-    /// Shorthand: `true` = enabled with defaults, `false` = disabled.
-    Bool(bool),
-}
-
-/// Empty options object for features that accept `{ ... }` but have no
-/// user-facing knobs yet. Fields are filled in by each feature's port
-/// sub-issue; this stub satisfies the schema shape requirement.
-///
-/// `deny_unknown_fields` is critical here — without it, the `FeatureToggle`
-/// untagged enum would accept ANY object as `Options(FeatureOptions {})`,
-/// silently turning `features.githubAlerts = { bogus: true }` into
-/// "enabled with defaults" instead of failing fast.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct FeatureOptions {}
-
-/// Options for the `githubAutolinks` feature. Requires `repo`.
-///
-/// TODO: fill in actual fields when the githubAutolinks feature is ported.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct GithubAutolinksConfig {
-    /// GitHub repository reference (`owner/repo`) used to build autolink URLs.
-    // Required by the githubAutolinks feature spec; value is set by users.
-    #[serde(default)]
-    pub repo: Option<String>,
-}
-
-/// Options stub for the `codeEnrichment` feature.
-///
-/// TODO: fill in actual fields when the codeEnrichment feature is ported.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct CodeEnrichmentConfig {}
-
-/// Options stub for the `tocExport` feature.
-///
-/// TODO: fill in actual fields when the tocExport feature is ported.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct TocExportConfig {}
-
-/// Options stub for the `imageDimensions` feature.
-///
-/// TODO: fill in actual fields when the imageDimensions feature is ported.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct ImageDimensionsConfig {}
-
-/// Options stub for the `linkValidation` feature.
-///
-/// TODO: fill in actual fields when the linkValidation feature is ported.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct LinkValidationConfig {}
-
-/// Options stub for the `transclude` feature.
-///
-/// TODO: fill in actual fields when the transclude feature is ported.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct TranscludeConfig {}
-
-/// Per-feature markdown pipeline configuration.
-///
-/// Each field corresponds to one pluggable markdown feature. All fields are
-/// optional; absent = feature disabled, behaviour unchanged from the
-/// pre-features build. Unknown keys are rejected at deserialization time
-/// (via `#[serde(deny_unknown_fields)]`) so a typo in `zfb.config.ts`
-/// surfaces as a clear error naming the unknown field rather than silently
-/// passing through with the feature disabled.
-///
-/// Mirrors `MarkdownFeaturesConfig` in `packages/zfb/src/config.ts`.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct MarkdownFeaturesConfig {
-    /// GitHub-style alert blocks (`> [!NOTE]`, `> [!WARNING]`, etc.).
-    #[serde(default)]
-    pub github_alerts: Option<FeatureToggle>,
-
-    /// Reading-time estimate injected into the document frontmatter.
-    #[serde(default)]
-    pub reading_time: Option<FeatureToggle>,
-
-    /// GitHub-style `owner/repo#123` and `SHA` autolinks. Requires `repo`.
-    #[serde(default)]
-    pub github_autolinks: Option<GithubAutolinksConfig>,
-
-    /// Code-block enrichment (copy button, language label, etc.).
-    #[serde(default)]
-    pub code_enrichment: Option<CodeEnrichmentConfig>,
-
-    /// Grouped code blocks rendered as tabs.
-    #[serde(default)]
-    pub code_tabs: Option<FeatureToggle>,
-
-    /// Ruby annotation support (`{base}^{ruby}` syntax).
-    #[serde(default)]
-    pub ruby: Option<FeatureToggle>,
-
-    /// Export the page TOC as structured data (e.g. for sidebar rendering).
-    #[serde(default)]
-    pub toc_export: Option<TocExportConfig>,
-
-    /// Auto-detect and inject `width`/`height` on `<img>` elements.
-    #[serde(default)]
-    pub image_dimensions: Option<ImageDimensionsConfig>,
-
-    /// Validate internal and external links at build time.
-    #[serde(default)]
-    pub link_validation: Option<LinkValidationConfig>,
-
-    /// Transclusion of other MDX files (`![[path]]` syntax).
-    #[serde(default)]
-    pub transclude: Option<TranscludeConfig>,
-
-    /// Framework admonitions preset (maps `:::note` etc. to components).
-    #[serde(default)]
-    pub admonitions_preset: Option<FeatureToggle>,
-
-    /// Mermaid diagram rendering.
-    #[serde(default)]
-    pub mermaid: Option<FeatureToggle>,
-
-    /// Lightbox / image-enlarge on click.
-    #[serde(default)]
-    pub image_enlarge: Option<FeatureToggle>,
-
-    /// Inline heading-marker TOC (e.g. `## Heading {#id}` syntax, TOC
-    /// rendered via a heading-marker plugin). Uses [`TocConfig`] shape.
-    #[serde(default)]
-    pub heading_marker_toc: Option<TocConfig>,
-}
+//
+// Canonical definitions live in `zfb-md-ast::features_config` so the
+// visitor-contract crate, `zfb-md-extras`, AND this user-facing `zfb`
+// crate all share a single source of truth — no parallel definitions,
+// no conversion bridges. Re-exported here so existing import paths
+// like `zfb::config::{MarkdownFeaturesConfig, FeatureToggle, ...}` and
+// `zfb::config::TocConfig` continue to resolve.
+pub use zfb_md_ast::{
+    CodeEnrichmentConfig, FeatureOptions, FeatureToggle, GithubAutolinksConfig,
+    ImageDimensionsConfig, LinkValidationConfig, MarkdownFeaturesConfig, TocConfig,
+    TocExportConfig, TranscludeConfig,
+};
 
 /// Options for the `rehype-external-links` port.
 ///
